backend/requirements.txt
django==4.1.7
requests==2.28.
djangorestframework==3.14.0
django-filter==22.1
djangorestframework-simplejwt==4.8.0
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
psycopg2-binary==2.9.5
Pillow==9.4.0